# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  entities.JSONUserInfo:
    properties:
      user:
        $ref: '#/definitions/entities.UserInfo'
    type: object
  entities.Recipe:
    properties:
      about:
        maxLength: 2500
        type: string
      complexitiy:
        enum:
        - 1
        - 2
        - 3
        maximum: 3
        minimum: 1
        type: integer
      created_at:
        type: string
      id:
        type: integer
      ingridients:
        maxLength: 1500
        type: string
      need_time:
        type: string
      photos_urls:
        type: string
      title:
        maxLength: 50
        minLength: 3
        type: string
      updated_at:
        type: string
    required:
    - about
    - complexitiy
    - ingridients
    - need_time
    - title
    type: object
...
  entities.UserInfo:
    properties:
      about:
        type: string
      created_at:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      liked_recipies:
        items:
          $ref: '#/definitions/entities.Recipe'
        type: array
      login:
        type: string
      recipies:
        items:
          $ref: '#/definitions/entities.Recipe'
        type: array
    type: object
  entities.UserLogin:
    properties:
      email:
        example: test@test.com
        type: string
      login:
        example: testuser
        type: string
      password:
        example: testpassword
        maxLength: 50
        minLength: 8
        type: string
    required:
    - password
    type: object
  entities.UserPasswordUpdate:
    properties:
      password:
        example: testpassword
        maxLength: 50
        minLength: 8
        type: string
    required:
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: RestAPI for recipe site
  title: RecipeSite
  version: "1.0"
paths:
  /auth/logout:
    post:
      description: Logout user
      operationId: Logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Logout
      tags:
      - auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Sign in user
      operationId: signin
      parameters:
      - description: User params
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entities.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: login
          schema:
            type: string
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Sign in
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Sign up user
      operationId: signup
      parameters:
      - description: User params
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entities.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Sign up
      tags:
      - auth
...
  /user/{login}:
    get:
      description: Get user info
      operationId: get user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.JSONUserInfo'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get user info
      tags:
      - user
    put:
      consumes:
      - multipart/form-data
      description: Update user
      operationId: update user
      parameters:
      - description: Icon
        in: formData
        name: icon
        type: file
      - in: formData
        maxLength: 1500
        name: about
        type: string
      - in: formData
        name: email
        type: string
      - in: formData
        maxLength: 20
        minLength: 3
        name: login
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: login
          schema:
            type: string
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update user
      tags:
      - user
  /user/{login}/password:
    put:
      consumes:
      - application/json
      description: Update user password
      operationId: update user password
      parameters:
      - description: User params
        in: body
        name: user
        schema:
          $ref: '#/definitions/entities.UserPasswordUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update user password
      tags:
      - user
...
swagger: "2.0"